feat(menu-service): Add complete project structure and initial CRUD logic

+++ auth-service/src/main/java/com/innsync/auth/util/DBUtility.java
@@ -1,7 +1,7 @@
 package com.innsync.auth.util;
 
 public class DBUtility {
-	public static String generateIdentifier(String businessName,Integer id) {
+	public static String generateIdentifier(String businessName,Long id) {
 		String cleanedName=businessName.toLowerCase().replaceAll("[^a-z0-9]", "_");
 		return "innsync_" + cleanedName + "_" + id;
 	}
